sort of added gradient to background but there are ugly black bars
