+++ src/components/background.module.css
@@ -1,0 +1,10 @@
+.background {
+  left:0;
+  top:0;
+  width: 100vw;
+  height: 100vh;
+  display: block;
+  position: fixed;
+  overflow: hidden;
+  z-index: -1;
+}
