+++ src/pages/_app.js
@@ -1,14 +1,16 @@
 import 'styles/fonts.css';
 import 'styles/globals.css';
 import NavBar from 'components/navbar';
+import { Background } from 'components/background';
 
 function MyApp({ Component, pageProps }) {
   return (
     <>
       <NavBar />
-      <div className="grid">
-        <Component {...pageProps} />
-      </div>
+      {/* <div className="grid"> */}
+      <Background />
+      <Component {...pageProps} />
+      {/* </div> */}
     </>
   );
 }
